+++ src/DependencyInjection/GosWebSocketExtension.php
@@ -58,21 +58,22 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
         $loader->load('services.yaml');
         $loader->load('aliases.yaml');
 
-        $configs = $this->processConfiguration($this->getConfiguration($configs, $container), $configs);
+        $config = $this->processConfiguration($this->getConfiguration($configs, $container), $configs);
 
         $container->registerForAutoconfiguration(PeriodicInterface::class)->addTag('gos_web_socket.periodic');
         $container->registerForAutoconfiguration(RpcInterface::class)->addTag('gos_web_socket.rpc');
         $container->registerForAutoconfiguration(ServerInterface::class)->addTag('gos_web_socket.server');
         $container->registerForAutoconfiguration(TopicInterface::class)->addTag('gos_web_socket.topic');
 
-        $container->setParameter('gos_web_socket.shared_config', $configs['shared_config']);
+        $container->setParameter('gos_web_socket.shared_config', $config['shared_config']);
 
-        $this->registerClientConfiguration($configs, $container);
-        $this->registerServerConfiguration($configs, $container);
-        $this->registerOriginsConfiguration($configs, $container);
-        $this->registerPingConfiguration($configs, $container);
-        $this->registerPushersConfiguration($configs, $container);
-        $this->registerWebsocketClientConfiguration($configs, $container);
+        $this->registerClientConfiguration($config, $container);
+        $this->registerServerConfiguration($config, $container);
+        $this->registerOriginsConfiguration($config, $container);
+        $this->registerBlockedIpAddressesConfiguration($config, $container);
+        $this->registerPingConfiguration($config, $container);
+        $this->registerPushersConfiguration($config, $container);
+        $this->registerWebsocketClientConfiguration($config, $container);
 
         $this->markAliasesDeprecated($container);
         $this->markServicesDeprecated($container);
@@ -130,20 +131,20 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
         }
     }
 
-    private function registerClientConfiguration(array $configs, ContainerBuilder $container): void
+    private function registerClientConfiguration(array $config, ContainerBuilder $container): void
     {
-        if (!isset($configs['client'])) {
+        if (!isset($config['client'])) {
             return;
         }
 
-        $container->setParameter('gos_web_socket.client.storage.ttl', $configs['client']['storage']['ttl']);
-        $container->setParameter('gos_web_socket.firewall', (array) $configs['client']['firewall']);
+        $container->setParameter('gos_web_socket.client.storage.ttl', $config['client']['storage']['ttl']);
+        $container->setParameter('gos_web_socket.firewall', (array) $config['client']['firewall']);
 
         // @deprecated to be removed in 4.0, parameter is unused
-        $container->setParameter('gos_web_socket.client.storage.prefix', $configs['client']['storage']['prefix']);
+        $container->setParameter('gos_web_socket.client.storage.prefix', $config['client']['storage']['prefix']);
 
-        if (isset($configs['client']['session_handler'])) {
-            $sessionHandler = ltrim($configs['client']['session_handler'], '@');
+        if (isset($config['client']['session_handler'])) {
+            $sessionHandler = ltrim($config['client']['session_handler'], '@');
 
             $container->getDefinition('gos_web_socket.server.builder')
                 ->addMethodCall('setSessionHandler', [new Reference($sessionHandler)]);
@@ -151,12 +152,12 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
             $container->setAlias('gos_web_socket.session_handler', $sessionHandler);
         }
 
-        if (isset($configs['client']['storage']['driver'])) {
-            $driverRef = ltrim($configs['client']['storage']['driver'], '@');
+        if (isset($config['client']['storage']['driver'])) {
+            $driverRef = ltrim($config['client']['storage']['driver'], '@');
             $storageDriver = $driverRef;
 
-            if (isset($configs['client']['storage']['decorator'])) {
-                $decoratorRef = ltrim($configs['client']['storage']['decorator'], '@');
+            if (isset($config['client']['storage']['decorator'])) {
+                $decoratorRef = ltrim($config['client']['storage']['decorator'], '@');
                 $container->getDefinition($decoratorRef)
                     ->addArgument(new Reference($driverRef));
 
@@ -171,38 +172,42 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
         }
     }
 
-    private function registerServerConfiguration(array $configs, ContainerBuilder $container): void
+    private function registerServerConfiguration(array $config, ContainerBuilder $container): void
     {
-        if (!isset($configs['server'])) {
+        if (!isset($config['server'])) {
             return;
         }
 
-        if (isset($configs['server']['port'])) {
-            $container->setParameter('gos_web_socket.server.port', $configs['server']['port']);
+        if (isset($config['server']['port'])) {
+            $container->setParameter('gos_web_socket.server.port', $config['server']['port']);
         }
 
-        if (isset($configs['server']['host'])) {
-            $container->setParameter('gos_web_socket.server.host', $configs['server']['host']);
+        if (isset($config['server']['host'])) {
+            $container->setParameter('gos_web_socket.server.host', $config['server']['host']);
         }
 
-        if (isset($configs['server']['origin_check'])) {
-            $container->setParameter('gos_web_socket.server.origin_check', $configs['server']['origin_check']);
+        if (isset($config['server']['origin_check'])) {
+            $container->setParameter('gos_web_socket.server.origin_check', $config['server']['origin_check']);
         }
 
-        if (isset($configs['server']['keepalive_ping'])) {
-            $container->setParameter('gos_web_socket.server.keepalive_ping', $configs['server']['keepalive_ping']);
+        if (isset($config['server']['ip_address_check'])) {
+            $container->setParameter('gos_web_socket.server.ip_address_check', $config['server']['ip_address_check']);
         }
 
-        if (isset($configs['server']['keepalive_interval'])) {
-            $container->setParameter('gos_web_socket.server.keepalive_interval', $configs['server']['keepalive_interval']);
+        if (isset($config['server']['keepalive_ping'])) {
+            $container->setParameter('gos_web_socket.server.keepalive_ping', $config['server']['keepalive_ping']);
         }
 
-        if (isset($configs['server']['router'])) {
+        if (isset($config['server']['keepalive_interval'])) {
+            $container->setParameter('gos_web_socket.server.keepalive_interval', $config['server']['keepalive_interval']);
+        }
+
+        if (isset($config['server']['router'])) {
             $routerConfig = [];
 
             // Adapt configuration based on the version of GosPubSubRouterBundle installed, if the XML loader is available the newer configuration structure is used
-            if (isset($configs['server']['router']['resources'])) {
-                foreach ($configs['server']['router']['resources'] as $resource) {
+            if (isset($config['server']['router']['resources'])) {
+                foreach ($config['server']['router']['resources'] as $resource) {
                     if (\is_array($resource)) {
                         $routerConfig[] = $resource;
                     } else {
@@ -218,25 +223,30 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
         }
     }
 
-    private function registerOriginsConfiguration(array $configs, ContainerBuilder $container): void
+    private function registerOriginsConfiguration(array $config, ContainerBuilder $container): void
     {
         $originsRegistryDef = $container->getDefinition('gos_web_socket.registry.origins');
 
-        foreach ($configs['origins'] as $origin) {
+        foreach ($config['origins'] as $origin) {
             $originsRegistryDef->addMethodCall('addOrigin', [$origin]);
         }
+    }
+
+    private function registerBlockedIpAddressesConfiguration(array $config, ContainerBuilder $container): void
+    {
+        $container->setParameter('gos_web_socket.blocked_ip_addresses', $config['blocked_ip_addresses']);
     }
 
     /**
      * @throws InvalidArgumentException if an unsupported ping service type is given
      */
-    private function registerPingConfiguration(array $configs, ContainerBuilder $container): void
+    private function registerPingConfiguration(array $config, ContainerBuilder $container): void
     {
-        if (!isset($configs['ping'])) {
+        if (!isset($config['ping'])) {
             return;
         }
 
-        foreach ((array) $configs['ping']['services'] as $pingService) {
+        foreach ((array) $config['ping']['services'] as $pingService) {
             switch ($pingService['type']) {
                 case Configuration::PING_SERVICE_TYPE_DOCTRINE:
                     $serviceRef = ltrim($pingService['name'], '@');
@@ -266,11 +276,11 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
         }
     }
 
-    private function registerPushersConfiguration(array $configs, ContainerBuilder $container): void
+    private function registerPushersConfiguration(array $config, ContainerBuilder $container): void
     {
         $usesSymfony51Api = method_exists(Definition::class, 'getDeprecation');
 
-        if (!isset($configs['pushers'])) {
+        if (!isset($config['pushers'])) {
             // Remove all of the pushers
             foreach (['gos_web_socket.pusher.amqp', 'gos_web_socket.pusher.wamp'] as $pusher) {
                 $container->removeDefinition($pusher);
@@ -283,9 +293,9 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
             return;
         }
 
-        if (isset($configs['pushers']['amqp']) && $this->isConfigEnabled($container, $configs['pushers']['amqp'])) {
+        if (isset($config['pushers']['amqp']) && $this->isConfigEnabled($container, $config['pushers']['amqp'])) {
             // Pull the 'enabled' field out of the pusher's config
-            $factoryConfig = $configs['pushers']['amqp'];
+            $factoryConfig = $config['pushers']['amqp'];
             unset($factoryConfig['enabled']);
 
             $connectionFactoryDef = new Definition(
@@ -321,9 +331,9 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
             $container->removeDefinition('gos_web_socket.pusher.amqp.push_handler');
         }
 
-        if (isset($configs['pushers']['wamp']) && $this->isConfigEnabled($container, $configs['pushers']['wamp'])) {
+        if (isset($config['pushers']['wamp']) && $this->isConfigEnabled($container, $config['pushers']['wamp'])) {
             // Pull the 'enabled' field out of the pusher's config
-            $factoryConfig = $configs['pushers']['wamp'];
+            $factoryConfig = $config['pushers']['wamp'];
             unset($factoryConfig['enabled']);
 
             $connectionFactoryDef = new Definition(
@@ -358,16 +368,16 @@ final class GosWebSocketExtension extends Extension implements PrependExtensionI
         }
     }
 
-    private function registerWebsocketClientConfiguration(array $configs, ContainerBuilder $container): void
+    private function registerWebsocketClientConfiguration(array $config, ContainerBuilder $container): void
     {
         $usesSymfony51Api = method_exists(Definition::class, 'getDeprecation');
 
-        if (!isset($configs['websocket_client']) || !$configs['websocket_client']['enabled']) {
+        if (!isset($config['websocket_client']) || !$config['websocket_client']['enabled']) {
             return;
         }
 
         // Pull the 'enabled' field out of the client's config
-        $factoryConfig = $configs['websocket_client'];
+        $factoryConfig = $config['websocket_client'];
         unset($factoryConfig['enabled']);
 
         $clientFactoryDef = new Definition(