+++ src/Server/App/ServerBuilder.php
@@ -3,6 +3,7 @@
 namespace Gos\Bundle\WebSocketBundle\Server\App;
 
 use Gos\Bundle\WebSocketBundle\Server\App\Registry\OriginRegistry;
+use Gos\Bundle\WebSocketBundle\Server\App\Stack\BlockedIpCheck;
 use Gos\Bundle\WebSocketBundle\Server\App\Stack\OriginCheck;
 use Gos\Bundle\WebSocketBundle\Server\App\Stack\WampConnectionPeriodicTimer;
 use Gos\Bundle\WebSocketBundle\Server\WampServer;
@@ -20,11 +21,16 @@ final class ServerBuilder implements ServerBuilderInterface
     private TopicManager $topicManager;
     private OriginRegistry $originRegistry;
     private EventDispatcherInterface $eventDispatcher;
-    private bool $originCheck = false;
-    private bool $keepalivePing = false;
-    private int $keepaliveInterval = 30;
+    private bool $originCheck;
+    private bool $keepalivePing;
+    private int $keepaliveInterval;
+    private bool $ipAddressCheck;
+    private array $blockedIpAddresses;
     private ?\SessionHandlerInterface $sessionHandler = null;
 
+    /**
+     * @param string[] $blockedIpAddresses
+     */
     public function __construct(
         LoopInterface $loop,
         TopicManager $topicManager,
@@ -32,7 +38,9 @@ final class ServerBuilder implements ServerBuilderInterface
         EventDispatcherInterface $eventDispatcher,
         bool $originCheck,
         bool $keepalivePing,
-        int $keepaliveInterval
+        int $keepaliveInterval,
+        bool $ipAddressCheck = false,
+        array $blockedIpAddresses = []
     ) {
         $this->loop = $loop;
         $this->topicManager = $topicManager;
@@ -41,6 +49,8 @@ final class ServerBuilder implements ServerBuilderInterface
         $this->originCheck = $originCheck;
         $this->keepalivePing = $keepalivePing;
         $this->keepaliveInterval = $keepaliveInterval;
+        $this->ipAddressCheck = $ipAddressCheck;
+        $this->blockedIpAddresses = $blockedIpAddresses;
     }
 
     public function buildMessageStack(): MessageComponentInterface
@@ -74,7 +84,17 @@ final class ServerBuilder implements ServerBuilderInterface
             );
         }
 
-        return new HttpServer($serverComponent);
+        $serverComponent = new HttpServer($serverComponent);
+
+        if ($this->ipAddressCheck && !empty($this->blockedIpAddresses)) {
+            $serverComponent = new BlockedIpCheck(
+                $serverComponent,
+                $this->eventDispatcher,
+                $this->blockedIpAddresses
+            );
+        }
+
+        return $serverComponent;
     }
 
     public function setSessionHandler(\SessionHandlerInterface $sessionHandler): void